+++ src/main/kotlin/omok/domain/omokboard/ColumnPosition.kt
@@ -1,0 +1,6 @@
+package omok.domain.omokboard
+
+@JvmInline
+value class ColumnPosition(
+    val value: Int,
+)
