+++ src/main/kotlin/omok/domain/omokboard/PointState.kt
@@ -1,0 +1,7 @@
+package omok.domain.omokboard
+
+enum class PointState {
+    OCCUPIED_BLACK,
+    OCCUPIED_WHITE,
+    EMPTY,
+}
