+++ src/main/kotlin/omok/domain/omokboard/Position.kt
@@ -1,0 +1,6 @@
+package omok.domain.omokboard
+
+data class Position(
+    val row: RowPosition,
+    val column: ColumnPosition,
+)
