+++ src/main/kotlin/omok/domain/omokboard/RowPosition.kt
@@ -1,0 +1,8 @@
+package omok.domain.omokboard
+
+@JvmInline
+value class RowPosition(
+    val value: Int,
+) {
+    override fun toString(): String = value.toString()
+}
