+++ src/main/kotlin/omok/domain/player/PlayerStone.kt
@@ -1,0 +1,8 @@
+package omok.domain.player
+
+import omok.domain.omokboard.Position
+
+data class PlayerStone(
+    val color: StoneColor,
+    val position: Position,
+)
