+++ src/main/kotlin/omok/domain/player/StoneColor.kt
@@ -1,0 +1,13 @@
+package omok.domain.player
+
+enum class StoneColor {
+    BLACK,
+    WHITE,
+    ;
+
+    fun reversed(): StoneColor =
+        when (this) {
+            BLACK -> WHITE
+            WHITE -> BLACK
+        }
+}
