+++ src/main/kotlin/omok/domain/rule/AlreadyExistStoneRule.kt
@@ -1,0 +1,17 @@
+package omok.domain.rule
+
+import omok.domain.omokboard.OmokBoard
+import omok.domain.omokboard.PointState
+import omok.domain.player.PlayerStone
+
+class AlreadyExistStoneRule : OmokRule {
+    override fun canPlace(
+        omokBoard: OmokBoard,
+        playerStone: PlayerStone,
+    ): PlaceResult =
+        if (omokBoard.find(playerStone.position)?.state == PointState.EMPTY) {
+            PlaceResult.Success.Progress(playerStone)
+        } else {
+            PlaceResult.Failure.AlreadyExistStone
+        }
+}
