+++ src/main/kotlin/omok/domain/rule/DrawRule.kt
@@ -1,0 +1,17 @@
+package omok.domain.rule
+
+import omok.domain.omokboard.OmokBoard
+import omok.domain.omokboard.PointState
+import omok.domain.player.PlayerStone
+
+class DrawRule : OmokRule {
+    override fun canPlace(
+        omokBoard: OmokBoard,
+        playerStone: PlayerStone,
+    ): PlaceResult =
+        if (omokBoard.value.values.all { it.state != PointState.EMPTY }) {
+            PlaceResult.Success.Finish(GameResult.DRAW)
+        } else {
+            PlaceResult.Success.Progress(playerStone)
+        }
+}
