+++ src/main/kotlin/omok/domain/rule/GameResult.kt
@@ -1,0 +1,7 @@
+package omok.domain.rule
+
+enum class GameResult {
+    WIN_BLACK,
+    WIN_WHITE,
+    DRAW,
+}
