--- src/main/kotlin/omok/model/domain/omokboard/PointState.kt
@@ -1,7 +1,0 @@
-package omok.model.domain.omokboard
-
-enum class PointState {
-    OCCUPIED_BLACK,
-    OCCUPIED_WHITE,
-    EMPTY,
-}
