--- src/main/kotlin/omok/model/domain/omokboard/Position.kt
@@ -1,6 +1,0 @@
-package omok.model.domain.omokboard
-
-data class Position(
-    val row: RowPosition,
-    val column: ColumnPosition,
-)
