--- src/main/kotlin/omok/model/domain/omokboard/RowPosition.kt
@@ -1,8 +1,0 @@
-package omok.model.domain.omokboard
-
-@JvmInline
-value class RowPosition(
-    val value: Int,
-) {
-    override fun toString(): String = value.toString()
-}
