--- src/main/kotlin/omok/model/domain/player/PlayerStone.kt
@@ -1,8 +1,0 @@
-package omok.model.domain.player
-
-import omok.model.domain.omokboard.Position
-
-data class PlayerStone(
-    val color: StoneColor,
-    val position: Position,
-)
