--- src/main/kotlin/omok/model/domain/player/StoneColor.kt
@@ -1,13 +1,0 @@
-package omok.model.domain.player
-
-enum class StoneColor {
-    BLACK,
-    WHITE,
-    ;
-
-    fun reversed(): StoneColor =
-        when (this) {
-            BLACK -> WHITE
-            WHITE -> BLACK
-        }
-}
