--- src/main/kotlin/omok/model/domain/rule/PlaceResult.kt
@@ -1,27 +1,0 @@
-package omok.model.domain.rule
-
-import omok.model.domain.player.PlayerStone
-
-sealed class PlaceResult {
-    sealed class Success : PlaceResult() {
-        data class Progress(
-            val lastStone: PlayerStone,
-        ) : Success()
-
-        data class Finish(
-            val winning: GameResult,
-        ) : Success()
-    }
-
-    sealed class Failure : PlaceResult() {
-        data object AlreadyExistStone : Failure()
-
-        data object InvalidPosition : Failure()
-
-        data object DoubleThreeViolation : Failure()
-
-        data object DoubleFourViolation : Failure()
-
-        data object OverlineViolation : Failure()
-    }
-}
