+++ src/test/kotlin/omok/Fixture.kt
@@ -1,8 +1,8 @@
 package omok
 
-import omok.model.domain.omokboard.ColumnPosition
-import omok.model.domain.omokboard.Position
-import omok.model.domain.omokboard.RowPosition
+import omok.domain.omokboard.ColumnPosition
+import omok.domain.omokboard.Position
+import omok.domain.omokboard.RowPosition
 
 val POSITION_ONE_ONE = Position(RowPosition(1), ColumnPosition(1))
